# 12_Backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ScottsTechX Commerce OS API",
    "version": "1.0.0",
    "description": "Trust-gated marketplace backend for the Uganda MVP slice.\n\nThis spec is the source of truth for clients (Flutter, web, admin). It mirrors\nthe implementation in `12_Backend/src/server.ts` and the migration-derived\nschema. Endpoints marked `STUB` return HTTP 501 with `X-Stub-Reason` header\nand do NOT perform any external I/O.\n\n## Authentication\n\nAll endpoints require `Authorization: Bearer ***` where the token is an\nHS256 JWT with the following claims:\n  - `iss`: must equal `scottstechx`\n  - `aud`: must equal `scottstechx-api`\n  - `sub`: user's UUID\n  - `role`: one of `buyer | driver | seller | admin`\n  - `exp`: present and not in the past\n\n## Idempotency\n\n`POST` and `PATCH` endpoints that change state require an `Idempotency-Key`\nheader of at least 8 characters. Replays with the same body return the\noriginal response. Replays with a different body under the same key return\nHTTP 409.\n\n## Error format\n\nAll errors return JSON: `{ error: string, message: string, details?: any }`.\nThe `error` field is a stable machine-readable code.",
    "license": { "name": "Proprietary" }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local dev" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SellerStats": {
        "type": "object",
        "required": ["sellerId","activeListings","totalListings","ordersToday","ordersThisWeek","revenueMinorToday","revenueMinorThisWeek","currency","averageRating","ratingCount"],
        "properties": {
          "sellerId": { "type": "string", "format": "uuid" },
          "activeListings": { "type": "integer" },
          "totalListings": { "type": "integer" },
          "ordersToday": { "type": "integer" },
          "ordersThisWeek": { "type": "integer" },
          "revenueMinorToday": { "type": "integer" },
          "revenueMinorThisWeek": { "type": "integer" },
          "currency": { "type": "string" },
          "averageRating": { "type": "number" },
          "ratingCount": { "type": "integer" }
        }
      },
      "CreateProductRequest": {
        "type": "object",
        "required": ["title","description","priceMinor","currency","stockQuantity"],
        "properties": {
          "title": { "type": "string", "maxLength": 80 },
          "description": { "type": "string", "maxLength": 500 },
          "priceMinor": { "type": "integer", "minimum": 0 },
          "currency": { "type": "string" },
          "stockQuantity": { "type": "integer", "minimum": 0 },
          "imageUrl": { "type": "string", "nullable": true }
        }
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "maxLength": 80 },
          "description": { "type": "string", "maxLength": 500 },
          "priceMinor": { "type": "integer", "minimum": 0 },
          "stockQuantity": { "type": "integer", "minimum": 0 },
          "imageUrl": { "type": "string", "nullable": true },
          "isActive": { "type": "boolean" }
        }
      },
...
      "SellerProfilePatch": {
        "type": "object",
        "properties": {
          "businessName": { "type": "string" },
          "businessDescription": { "type": "string", "nullable": true },
          "address": { "type": "string", "nullable": true },
          "lat": { "type": "number", "nullable": true },
          "lng": { "type": "number", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "bannerUrl": { "type": "string", "nullable": true },
          "opensAt": { "type": "string", "nullable": true },
          "closesAt": { "type": "string", "nullable": true }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/seller/profile": {
      "get": {
        "tags": ["marketplace"], "summary": "Get my seller profile (seller only)",
        "responses": {
          "200": { "description": "Profile", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not a seller" }
        }
      },
      "patch": {
        "tags": ["marketplace"], "summary": "Update my seller profile",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SellerProfilePatch" } } } },
        "responses": {
          "200": { "description": "Updated" },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not a seller" },
          "404": { "description": "No fields to update" }
        }
      }
    },
    "/api/v1/seller/inventory": {
      "get": {
        "tags": ["marketplace"], "summary": "List my inventory (seller only)",
        "responses": {
          "200": { "description": "Inventory list" },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not a seller" }
        }
      },
      "post": {
        "tags": ["marketplace"], "summary": "Create product (seller only)",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 8 } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProductRequest" } } } },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error or missing Idempotency-Key" },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not a seller" }
        }
      }
    },
    "/api/v1/seller/inventory/{productId}": {
      "patch": {
        "tags": ["marketplace"], "summary": "Patch product (seller only)",
        "parameters": [
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 8 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProductRequest" } } } },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "No fields to update" },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not the product's seller" },
          "404": { "description": "Product not found" }
        }
      },
      "delete": {
        "tags": ["marketplace"], "summary": "Soft-delete product (seller only)",
        "parameters": [
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 8 } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not the product's seller" },
          "404": { "description": "Product not found" }
        }
      }
    },
    "/api/v1/seller/stats": {
      "get": {
        "tags": ["marketplace"], "summary": "Dashboard stats for the signed-in seller",
        "responses": {
          "200": { "description": "Stats", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SellerStats" } } } },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not a seller" }
        }
      }
    },
    "/api/v1/seller/orders": {
      "get": {
        "tags": ["marketplace"], "summary": "Orders for the signed-in seller",
        "parameters": [{ "name": "status", "in": "query", "required": false, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Orders" },
          "401": { "description": "Missing or invalid JWT" },
          "403": { "description": "Caller is not a seller" }
        }
      }
    },
...
  }
}